Moved the user APIHandlers into a user map
