+++ java/src/api/handlers/user/APILogoutHandler.java
@@ -1,4 +1,4 @@
-package api.handlers;
+package api.handlers.user;
 
 import static api.helpers.EntityContentHelper.checkAndGetEntityContent;
 import static api.helpers.RequestMethodHelper.checkRequestMethod;
@@ -7,29 +7,26 @@ import api.exceptions.APIBadRequestException;
 import org.apache.http.HttpRequest;
 import org.json.JSONObject;
 
-public class APILoginHandler {
+public class APILogoutHandler {
 
-  public static String handleLoginRequest(HttpRequest httpRequest) {
+  public static String handleLogoutRequest(HttpRequest httpRequest){
     checkRequestMethod("POST", httpRequest);
 
     String requestContent = checkAndGetEntityContent(httpRequest);
 
     JSONObject jsonObject = new JSONObject(requestContent);
 
-    if (!jsonObject.has("username") || !jsonObject.has("password")) {
-      throw new APIBadRequestException("Login data not complete");
+    if (!jsonObject.has("username") || !jsonObject.has("token")){
+      throw new APIBadRequestException("Logout data not complete");
     }
 
     String username = jsonObject.getString("username");
     String password = jsonObject.getString("password");
 
-    JSONObject loginResponse = new JSONObject();
-    loginResponse.put("username", username);
-    loginResponse.put("token", password);
+    JSONObject logoutResponse = new JSONObject();
+    logoutResponse.put("logged_out", "true");
 
-    return loginResponse.toString();
+    return logoutResponse.toString();
   }
-
-
 
 }
